+++ com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/bundle/HTML5Bundler.java
@@ -145,15 +145,21 @@ public class HTML5Bundler implements IBundler {
     }
 
     @Override
-    public void bundleApplication(Project project, File bundleDirectory)
+    public void bundleApplication(Project project, File bundleDirectory, ICanceled canceled)
             throws IOException, CompileExceptionError {
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         // Collect bundle/package resources to be included in bundle directory
         Map<String, IResource> bundleResources = ExtenderUtil.collectResources(project, Platform.JsWeb);
 
         BobProjectProperties projectProperties = project.getProjectProperties();
         BobProjectProperties metaProperties = getPropertiesMeta();
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         Boolean localLaunch = project.option("local-launch", "false").equals("true");
         final String variant = project.option("variant", Bob.VARIANT_RELEASE);
         String title = projectProperties.getStringValue("project", "title", "Unnamed");
@@ -163,6 +169,9 @@ public class HTML5Bundler implements IBundler {
         List<File> binsAsmjs = null;
         List<File> binsWasm = null;
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         // asmjs binaries
         {
             Platform targetPlatform = Platform.JsWeb;
@@ -176,6 +185,9 @@ public class HTML5Bundler implements IBundler {
             ;
         }
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         // wasm binaries
         {
             Platform targetPlatform = Platform.WasmWeb;
@@ -188,6 +200,9 @@ public class HTML5Bundler implements IBundler {
             }
         }
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         File projectRoot = new File(project.getRootDirectory());
         URL splashImage = getResource(projectProperties, projectRoot, "html5", "splash_image", "splash_image.png");
         String version = projectProperties.getStringValue("project", "version", "0.0");
@@ -205,6 +220,9 @@ public class HTML5Bundler implements IBundler {
             customHeapSize = 256*1024*1024; // Same value as engine is compiled with; 268435456
         }
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         Map<String, Object> infoData = new HashMap<String, Object>();
         infoData.put("DEFOLD_ENGINE", enginePrefix + ".js");
         infoData.put("DEFOLD_BINARY_PREFIX", enginePrefix);
@@ -231,6 +249,9 @@ public class HTML5Bundler implements IBundler {
             infoData.put("DEFOLD_ARCHIVE_LOCATION_SUFFIX", projectProperties.getStringValue("html5", "archive_location_suffix", ""));
         }
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         String devInit = "";
         String devHead = "";
         String inlineHtml = "";
@@ -244,16 +265,25 @@ public class HTML5Bundler implements IBundler {
         infoData.put("DEFOLD_DEV_HEAD", devHead);
         infoData.put("DEFOLD_DEV_INLINE", inlineHtml);
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         FileUtils.deleteDirectory(appDir);
         File splitDir = new File(appDir, SplitFileDir);
         splitDir.mkdirs();
         createSplitFiles(buildDir, splitDir);
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         // Copy bundle resources into bundle directory
         ExtenderUtil.writeResourcesToDirectory(bundleResources, appDir);
 
         // Copy engine binaries
         for (File bin : binsAsmjs) {
+            if(canceled.isCanceled()) {
+                return;
+            }
             String binExtension = FilenameUtils.getExtension(bin.getAbsolutePath());
             if (binExtension.equals("js")) {
                 FileUtils.copyFile(bin, new File(appDir, enginePrefix + "_asmjs.js"));
@@ -263,6 +293,9 @@ public class HTML5Bundler implements IBundler {
         }
 
         for (File bin : binsWasm) {
+            if(canceled.isCanceled()) {
+                return;
+            }
             String binExtension = FilenameUtils.getExtension(bin.getAbsolutePath());
             if (binExtension.equals("js")) {
                 FileUtils.copyFile(bin, new File(appDir, enginePrefix + "_wasm.js"));
@@ -273,6 +306,9 @@ public class HTML5Bundler implements IBundler {
             }
         }
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         // Flash audio swf
         FileUtils.copyFile(new File(Bob.getLibExecPath("js-web/defold_sound.swf")), new File(appDir, "defold_sound.swf"));
