+++ com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/bundle/LinuxBundler.java
@@ -39,7 +39,7 @@ public class LinuxBundler implements IBundler {
     }
 
     @Override
-    public void bundleApplication(Project project, File bundleDir)
+    public void bundleApplication(Project project, File bundleDir, ICanceled canceled)
             throws IOException, CompileExceptionError {
 
         final Platform platform = Platform.X86_64Linux;
@@ -48,6 +48,10 @@ public class LinuxBundler implements IBundler {
         String exeName = BundleHelper.projectNameToBinaryName(title);
         File appDir = new File(bundleDir, title);
 
+        if(canceled.isCanceled()) {
+            return;
+        }
+
         FileUtils.deleteDirectory(appDir);
         appDir.mkdirs();
 
@@ -55,17 +59,33 @@ public class LinuxBundler implements IBundler {
         final List<String> extensionFolders = ExtenderUtil.getExtensionFolders(project);
         final boolean hasExtensions = !extensionFolders.isEmpty();
 
+        if(canceled.isCanceled()) {
+            return;
+        }
+
         bundleApplicationForPlatform(platform, project, appDir, exeName);
 
         File buildDir = new File(project.getRootDirectory(), project.getBuildDirectory());
+
+        if(canceled.isCanceled()) {
+            return;
+        }
 
         // Copy archive and game.projectc
         for (String name : Arrays.asList("game.projectc", "game.arci", "game.arcd", "game.dmanifest", "game.public.der")) {
             FileUtils.copyFile(new File(buildDir, name), new File(appDir, name));
         }
 
+        if(canceled.isCanceled()) {
+            return;
+        }
+
         // Collect bundle/package resources to be included in bundle directory
         Map<String, IResource> bundleResources = ExtenderUtil.collectResources(project, platform);
+
+        if(canceled.isCanceled()) {
+            return;
+        }
 
         // Copy bundle resources into bundle directory
         ExtenderUtil.writeResourcesToDirectory(bundleResources, appDir);
